check `401 UNAUTHORIZED`

+++ domain/shipment/shipment_router.py
@@ -3,7 +3,7 @@ from starlette import status
 
 from database import create_session
 from domain.shipment import shipment_schema, shipment_crud
-from domain.shipment.shipment_crud import delete_shipment
+from domain.shipment.shipment_crud import delete_shipment, convert_to_model
 
 router = APIRouter(prefix="/api/shipment")
 
@@ -12,9 +12,12 @@ router = APIRouter(prefix="/api/shipment")
 def shipping_order(
         shipment_in: shipment_schema.ShipmentIn | list[shipment_schema.ShipmentIn],
         session=Depends(create_session),
-        access_token: str = Header(convert_underscores=False)
+        access_token: str | None = Header(convert_underscores=False, default=None)
 ):
     try:
+        if access_token is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+
         shipment_crud.create_shipment(
             session=session,
             orders=shipment_in,
@@ -32,9 +35,12 @@ def shipping_order(
 def shipping_order(
         shipment_id: int,
         session=Depends(create_session),
-        access_token: str = Header(convert_underscores=False)
+        access_token: str | None = Header(convert_underscores=False, default=None)
 ):
     try:
+        if access_token is None:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+
         delete_shipment(session, access_token, shipment_id)
         return {"ok": True}
